Move findPriceTableByProductId under product directory
