+++ src/main/java/com/flab/bbt/deal/repository/DealRepository.java
@@ -1,7 +1,6 @@
 package com.flab.bbt.deal.repository;
 
 import com.flab.bbt.deal.domain.Deal;
-import com.flab.bbt.product.domain.PriceTable;
 import java.util.Optional;
 
 public interface DealRepository {
@@ -17,6 +16,4 @@ public interface DealRepository {
     int updateParticipantCountById(int updatedCount, Long id);
 
     int updateExpiredDeals();
-
-    Optional<PriceTable> findPriceTableByProductId(Long productId);
 }
