+++ src/main/java/com/flab/bbt/deal/repository/mybatis/DealMapper.java
@@ -2,7 +2,6 @@ package com.flab.bbt.deal.repository.mybatis;
 
 import com.flab.bbt.deal.domain.Deal;
 import com.flab.bbt.deal.domain.DealStatus;
-import com.flab.bbt.product.domain.PriceTable;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import org.apache.ibatis.annotations.Mapper;
@@ -25,5 +24,5 @@ public interface DealMapper {
 
     Optional<Deal> findByIdForUpdate(Long id);
 
-    Optional<PriceTable> findPriceTableByProductId(Long productId, LocalDateTime currentTime);
+
 }
